+++ react/ProductAssemblyOptionItemQuantity.tsx
@@ -1,0 +1,2 @@
+import ProductAssemblyOptionItemQuantity from './components/ProductAssemblyOptions/ProductAssemblyOptionItemQuantity'
+export default ProductAssemblyOptionItemQuantity