+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support recursive assembly options.
+- New components:
+	- `assembly-option-item-image`
+	- `assembly-option-item-quantity-selector`
+	- `assembly-option-item-name`
+	- `assembly-option-item-price`
+	- `assembly-option-item-customize`
+	- `assembly-option-item-children-description`
 
 ## [2.0.7] - 2019-04-05
 ### Changed